+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Update minimal node version to 14 [\#2034](https://github.com/reimagined/resolve/pull/2034)
 
+**Documentation:**
+
+- Describe API for fundamental entities of reSolve [\#2032](https://github.com/reimagined/resolve/pull/2032)
+
 ## [V0.31.10](https://github.com/reimagined/resolve/tree/V0.31.10) (2021-09-08)
 
 [Full Changelog](https://github.com/reimagined/resolve/compare/V0.31.9...V0.31.10)