+++ CHANGELOG.md
@@ -4,9 +4,17 @@
 
 [Full Changelog](https://github.com/reimagined/resolve/compare/V0.31.10...HEAD)
 
+**New features:**
+
+- Resolve gate [\#2040](https://github.com/reimagined/resolve/pull/2040)
+
 **Breaking changes:**
 
 - Update minimal node version to 14 [\#2034](https://github.com/reimagined/resolve/pull/2034)
+
+**Bug fixes:**
+
+- Timestamps from CommandResult and Event are different [\#1994](https://github.com/reimagined/resolve/issues/1994)
 
 **Documentation:**