+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 **Bug fixes:**
 
 - Timestamps from CommandResult and Event are different [\#1994](https://github.com/reimagined/resolve/issues/1994)
+- The Replicator read model takes a lot of resources during replication of big event store \(5kk events\) [\#1949](https://github.com/reimagined/resolve/issues/1949)
 
 **Documentation:**