+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 
 **Other changes:**
 
+- Next step on long builds [\#2056](https://github.com/reimagined/resolve/pull/2056)
+- Bump resolve-cloud-common version [\#2054](https://github.com/reimagined/resolve/pull/2054)
 - Update invocation type for send notifications [\#2053](https://github.com/reimagined/resolve/pull/2053)
 
 ## [V0.31.10](https://github.com/reimagined/resolve/tree/V0.31.10) (2021-09-08)